Function 5
Function 5 is a 4D chemical yield optimisation task, where the goal is to find the combination of inputs that produces the highest output. Unlike other functions, it is expected to be unimodal, meaning there should be a single global maximum rather than many local peaks. This makes the search more straightforward: after a brief exploratory phase, the best strategy is to switch to exploitation and refine around the most promising region. In Week 1, I therefore applied a more exploitative acquisition function (PI with a low exploration term), as this aligns well with the unimodal nature of the function.